remove redundant members
